project/requirements.txt
fastapi
uvicorn
python-dotenv
requests
Pillow
google-generativeai
cloudinary